+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## [0.21.0] - 2026-07-06
+
+### 🚀 Features
+
+- *(types)* Add rwaMarketType to Market ([#111](https://github.com/backpack-exchange/bpx-api-client/issues/111))
+## [0.20.6] - 2026-07-01
+
+### 🐛 Bug Fixes
+
+- *(types)* Skip None fields in FillsHistoryParams query serialization ([#109](https://github.com/backpack-exchange/bpx-api-client/issues/109))
+## [0.20.5] - 2026-06-30
+
+### 🚀 Features
+
+- *(types)* Expose market order book state and add unknown enum fallbacks ([#108](https://github.com/backpack-exchange/bpx-api-client/issues/108))
+## [0.20.4] - 2026-06-18
+
+### 🚀 Features
+
+- Add dependabot ([#91](https://github.com/backpack-exchange/bpx-api-client/issues/91))
+
+### ⚙️ Miscellaneous Tasks
+
+- *(types)* Remove serde and strum rename attributes for blockchain ([#107](https://github.com/backpack-exchange/bpx-api-client/issues/107))
+- Bump reqwest, tokio-tungstenite, serde_qs and strum versions ([#104](https://github.com/backpack-exchange/bpx-api-client/issues/104))
+## [0.20.3] - 2026-06-07
+
+### 🐛 Bug Fixes
+
+- Updates to the securities endpoint ([#106](https://github.com/backpack-exchange/bpx-api-client/issues/106))
 ## [0.20.1] - 2026-04-21
 
 ### 🚀 Features